Add persistent hash
only saves valid entries from the hash file, so the final size of the persistent file is not necessarily the same as the size of the hash.
Also, when loading, the current hash will not be resized, it will simply load the file entries into the current hash without resizing.